crawler/requirement.txt
beautifulsoup4==4.7.1
click==7.0
schedule==0.6.0
elasticsearch==5.5.3
lxml==4.3.4
urllib3==1.23
fake-useragent==0.1.11
selenium==3.141.0
requests==2.21.0
pytz==2019.1 
PyMySQL==0.9.3
Flask==1.0.3
